fix: move constant
