+++ packages/cli/src/react/jsx/utils/constants.ts
@@ -51,6 +51,3 @@ export const GT_ATTRIBUTES = [
 
 // Data attribute prefix injected by build tools
 export const DATA_ATTR_PREFIX = 'data-' as const;
-
-// Number of source code lines to capture above and below a translation site
-export const SURROUNDING_LINE_COUNT = 5;
